requirements.txt
dspy
openai
unstructured
streamlit
python-dotenv
requests
dspy-ai
PyPDF2
streamlit-sortables
markdown2 
reportlab
chromadb
PyMuPDF
sentence-transformers